fix: make CSV export prompt for a file download
This is a quick and dirty fix to make the export feature prompt the
user to save a file. It works in Chrome and Firefox and some versions
of IE. On Safari it will only open the content in the same window,
which is moderately annoying but not the end of the world.

This isn't a perfect solution but it appears to work well enough for
the short term. Longer term we'll want to integrate better support
for generating the CSV on the server side where it can be done in
a streaming environment without the need for a data URI.

This also uncovered a bug with the export feature where it was
generating invalid CSV.

This closes work item 43894.

Signed-off-by: Patrick Wagstrom <pwagstro@us.ibm.com>