docs: added information on password and upgrading
On the recommendation from Allen I took the password stuff from the general
steps at
https://github.com/watson-developer-cloud/questions-with-classifier-ega#stage-3-train-the-classifier.
The upgrade stuff is actually pretty easy with the Deploy to Bluemix button
because it will just reuse the existing service instances. We just need to
make sure the users know how to delete their existing services.

Signed-off-by: Patrick Wagstrom <pwagstro@us.ibm.com>