Fix: Adding socketio to both front and backend to handle all database calls (create, update, and delete). Creating a content service to handle import on the backend.
This very large update replaces a lot of functionality that was handled on the front end and moves it to the back end. All calls on the front end are done at once and the individual responses come back through the socket to handle both errors and successes. Import is changed to uploading a file to the backend for processing. There are still issues with adding in multer appropriately and making the database calls to ensure that only the diff is added.

bug#44047, bug#44046, story#44077, story#44439