fix(bulk-delete): have bulk delete follow async API conventions
A framework has been established to handle bulk/long-running operations.  The Bulk Delete functionality was not tied into this, and was incorrectly writing errors to the response stream after it was closed.  This fix establishes a common framework for this long running jobs (reducing some copy/paste) as well as brings the Bulk Delete functionality in the fold.

closes story 45215

Signed-Off-by: Andy Stoneberg <astonebe@us.ibm.com>