docs(services): how to use your own instance
Added the file Using-Your-Own-Service.md to the file. This file, created by
me and reviewed by Allen Dean describes the process for binding a Natural
Language Classifier Toolkit to a pre-existing service instance. Also added
a small section in the main README.md to indicate that's an option.

This closes work item 43909.

Signed-off-by: Patrick Wagstrom <pwagstro@us.ibm.com>