defect 46502: fall through to any bound classifier
This introduces a change in the way that the system connects to classifier
instances. Prior to this change it would connect to whatever service was
defined in CLASSIFIER_SERVICE_NAME. If that variable wasn't defined it would
fall through to ibmwatson-nlc-classifier. If there wasn't a service instance
with that name bound to the application then the application would just fail.

This change makes it so after falling through the previous conditions the
application will pick the first instance of a natural_language_classifier
service bound to the application and use those credentials.

It also adds a little bit of logging indicating that it's going to crash
because of the lack of a bound classifier.

Signed-off-by: Patrick Wagstrom <pwagstro@us.ibm.com>