defect 46502: unit test for multiple bound NLCs
This adds in another unit tests for a case I thought of but wasn't testing in
the previous tests. Basically, what happens if you have two different instances
of NLC bound to the tooling, we really need it to use a consistent classifier
service instance, so we default to the first. This new test covers that case.

Signed-off-by: Patrick Wagstrom <pwagstro@us.ibm.com>